Random events in double battles can happen one at a time and fixed SelectedForView to work with battle clones

+++ src/Types/Random.hpp
@@ -6,5 +6,5 @@ namespace pokesim::types {
 using probability = float;
 using rngState = std::uint64_t;
 using rngResult = std::uint32_t;
-using percentChance = std::uint32_t;
+using percentChance = std::uint8_t;
 }  // namespace pokesim::types
